requirements.txt
# Hootcam Motion – NUC app (motion detection + recording from RTSP)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
bcrypt>=4.0.0
pydantic>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
starlette>=0.27.0
